add edit mudule file
